+++ M2/Macaulay2/d/Makefile.files.in
@@ -243,8 +243,8 @@ engine.o :						\
 	@srcdir@/../e/monoid.hpp			\
 	@srcdir@/../e/imonorder.hpp			\
 	@srcdir@/../e/interface/monomial-ordering.h	\
-	@srcdir@/../e/freemod.hpp			\
-	@srcdir@/../e/schorder.hpp			\
+	@srcdir@/../e/free-modules/freemod.hpp		\
+	@srcdir@/../e/free-modules/schorder.hpp		\
 	@srcdir@/../e/monideal.hpp			\
 	@srcdir@/../e/int-bag.hpp			\
 	@srcdir@/../e/rings/polyring.hpp			\
