+++ M2/Macaulay2/e/f4/f4-m2-interface.cpp
@@ -3,7 +3,7 @@
 #include "f4/f4-m2-interface.hpp"
 #include "f4/moninfo.hpp"              // for monomial_word, MonomialInfo
 #include "f4/ntuple-monomial.hpp"      // for ntuple_word
-#include "freemod.hpp"                 // for FreeModule
+#include "free-modules/freemod.hpp"                 // for FreeModule
 #include "gbring.hpp"                  // for gbvector, GBRing
 #include "interface/mutable-matrix.h"  // for IM2_MutableMatrix_make
 #include "basic-mutable-matrices/mat.hpp"                     // for MutableMatrix
