+++ M2/Macaulay2/e/free-modules/freemod.cpp
@@ -1,7 +1,7 @@
 // Copyright 1996  Michael E. Stillman
 
 #include "util.hpp"
-#include "freemod.hpp"
+#include "free-modules/freemod.hpp"
 #include "comb.hpp"
 #include "rings/polyring.hpp"
 #include "matrix.hpp"
