+++ M2/Macaulay2/e/free-modules/freemod.hpp
