+++ M2/Macaulay2/e/interface/cra.cpp
@@ -2,7 +2,7 @@
 
 #include "cra.hpp"
 #include "error.h"
-#include "freemod.hpp"
+#include "free-modules/freemod.hpp"
 #include "matrix-con.hpp"
 #include "matrix.hpp"
 #include "monoid.hpp"
