+++ M2/Macaulay2/e/interface/freemodule.cpp
@@ -5,7 +5,7 @@
 #include "buffer.hpp"
 #include "error.h"
 #include "exceptions.hpp"
-#include "freemod.hpp"
+#include "free-modules/freemod.hpp"
 #include "monoid.hpp"
 #include "newdelete.hpp"
 #include "rings/ring.hpp"
