+++ M2/Macaulay2/e/matrix.cpp
@@ -18,7 +18,7 @@
 #include "assprime.hpp"
 #include "monideal.hpp"
 #include "relem.hpp"
-#include "freemod.hpp"
+#include "free-modules/freemod.hpp"
 #include "util.hpp"
 
 #include "exptable.h"
